- Change AdvancedExamples folder to Advanced Features - Add DataVisualization folder - Add ExtraModels Folder - Add StartingModels folder - rename Getting-Started.ipynb to StartHere.ipynb - Move files and Notebooks to appropriate folders - edits to README.md to reflect changes
