+++ examples/ExtraModels/degradation_example.py
