+++ examples/ExtraModels/tyson_oscillator.py
