+++ examples/README.md
@@ -3,8 +3,12 @@ Examples of using GillesPy2
 
 The files in this directory are runnable Python examples of using [GillesPy2](https://github.com/gillesPy2/GillesPy2) to perform stochastic simulation.  In terms of biology, they are overly simplistic and do not capture the real-life complexity of the process being modeled &ndash; the aim is not biological realism but rather to illustrate basic usage of [GillesPy2](https://github.com/gillesPy2/GillesPy2).
 
-* [Getting Started](Getting-Started.ipynb) &ndash; a [Jupyter Notebook](https://jupyter-notebook.readthedocs.io/en/stable/) demonstrating the use of GillesPy2 on a simple Michaelis-Menten model
-* [Basic Examples](BasicExamples) &ndash; numerous basic examples, in the form of standalone Python files and [Jupyter Notebook](https://jupyter-notebook.readthedocs.io/en/stable/)
-* [Advanced Examples](AdvancedExamples) &ndash; examples illustrating more complicated models or features
-* [SBML]() &ndash; examples and tests of reading models in [SBML](http://sbml.org) format
-* [Speed Comparisons](SpeedComparisons) &ndash; time comparison of different solvers
+* [Start Here](StartHere.ipynb) &ndash; a [Jupyter Notebook](https://jupyter-notebook.readthedocs.io/en/stable/) demonstrating the use of GillesPy2 on a simple Michaelis-Menten model
+
+* [Starting Models](StartingModels) &ndash; numerous basic files and [Jupyter Notebook](https://jupyter-notebook.readthedocs.io/en/stable/), showcasing all GillesPy2 solvers on two data sets, as well as speed comparisons  
+
+* [Advanced Features](AdvancedFeatures) &ndash; examples illustrating more complicated models or features, including how to import SBML files
+
+* [Extra Models](ExtraModels) &ndash; various extra models and files
+
+* [Data Visualization](DataVisualization) &ndash; contains files and [Jupyter Notebook](https://jupyter-notebook.readthedocs.io/en/stable/) demonstrating various methods of visualizing data with GillesPy2
