+++ examples/StartingModels/MichaelisMenten/Michaelis-Menten.py
@@ -1,0 +1,31 @@
+from gillespy2.core import Model, Species, Reaction, Parameter
+import numpy as np
+
+
+class MichaelisMenten(Model):
+    def __init__(self, parameter_values=None):
+        # initialize Model
+        Model.__init__(self, name="Michaelis_Menten")
+
+        # parameters
+        rate1 = Parameter(name='rate1', expression=0.0017)
+        rate2 = Parameter(name='rate2', expression=0.5)
+        rate3 = Parameter(name='rate3', expression=0.1)
+        self.add_parameter([rate1, rate2, rate3])
+
+        # Species
+        A = Species(name='A', initial_value=301)
+        B = Species(name='B', initial_value=120)
+        C = Species(name='C', initial_value=0)
+        D = Species(name='D', initial_value=0)
+        self.add_species([A, B, C, D])
+
+        # reactions
+        r1 = Reaction(name="r1", reactants={A: 1, B: 1}, products={C: 1}, rate=rate1)
+
+        r2 = Reaction(name="r2", reactants={C: 1}, products={A: 1, B: 1}, rate=rate2)
+
+        r3 = Reaction(name="r3", reactants={C: 1}, products={B: 1, D: 1}, rate=rate3)
+        self.add_reaction([r1, r2, r3])
+        self.timespan(np.linspace(0, 100, 101))
+
