+++ examples/StartingModels/SpeedComparisons/TimeComparisonOfSolvers.pdf
