+++ examples/StartingModels/VilarOscillator/VilarOscillator.py
